vpn-bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
qrcode==7.4.2
pillow==10.1.0
